Add probe intelligence + decrease reset interval for missing data check
